增加了车位模块的列表显示，车辆模块的列表显示

+++ src/main/java/com/gec/myparking/service/UserService.java
@@ -135,6 +135,17 @@ public class UserService {
     }
 
     /**
+     * 获得指定id用户
+     * @param id
+     * @return
+     */
+    public User getUserById(Integer id)
+    {
+        return userMapper.selectByPrimaryKey(id);
+    }
+
+
+    /**
      * 获取指定页的用户列表
      * @param page
      * @param limit
@@ -157,7 +168,7 @@ public class UserService {
     }
 
     public void deleteUser(Integer id) {
-        if (id !=null)
+        if (id !=null && userMapper.selectByPrimaryKey(id)!=null)
             userMapper.deleteByPrimaryKey(id);
         else
             throw  new NullPointerException();
